+++ src/crawling/crawling.module.ts
@@ -2,12 +2,13 @@ import { CollectionModule } from '../collection/collection.module';
 import { Module } from '@nestjs/common';
 
 import { RabbitMQModule } from '@golevelup/nestjs-rabbitmq';
-import { CrawlingService } from './crawling.service';
+import { RabbitMqService } from './rabbitmq.service';
 import { SlackAlertService } from './slack-alert.service';
 import { CrawlingController } from './crawling.controller';
 import { FirebaseService } from './firebase.service';
 import { UserModule } from 'src/user/user.module';
 import { PlaceModule } from 'src/place/place.module';
+import { CrawlingService } from './crawling.service';
 
 @Module({
   imports: [
@@ -40,7 +41,12 @@ import { PlaceModule } from 'src/place/place.module';
     PlaceModule,
     UserModule,
   ],
-  providers: [CrawlingService, SlackAlertService, FirebaseService],
+  providers: [
+    RabbitMqService,
+    SlackAlertService,
+    FirebaseService,
+    CrawlingService,
+  ],
   controllers: [CrawlingController],
 })
 export class CrawlingModule {}
